Move color conversion utilities out of renderers
