+++ docs/research/runtime_game_loop_modularization.md
@@ -18,8 +18,8 @@ that runtime services and rendering helpers stay focused and easier to reuse.
   `src/heart/environment.py` is a compatibility shim that re-exports the runtime
   symbols for legacy imports.
 - HSV/BGR conversion helpers and the conversion cache moved to
-  `src/heart/renderers/color_conversion.py`, keeping rendering utilities scoped
-  to the renderer layer instead of the runtime loop.
+  `src/heart/utilities/color_conversion.py`, keeping the conversion helpers
+  available as shared utilities outside the renderer layer.
 - Modules that configure or record the runtime loop now import
   `heart.runtime.game_loop` directly, which makes the entry points explicit and
   reduces the dependency surface of the environment shim.
@@ -27,7 +27,7 @@ that runtime services and rendering helpers stay focused and easier to reuse.
 ## References
 
 - `src/heart/runtime/game_loop.py`
-- `src/heart/renderers/color_conversion.py`
+- `src/heart/utilities/color_conversion.py`
 - `src/heart/environment.py`
 - `src/heart/display/recorder.py`
 - `src/heart/loop.py`
